+++ apps/imadm/src/adm_app_version_handler.erl
@@ -31,7 +31,7 @@ init(Req0, State0) ->
     Method = cowboy_req:method(Req0),
     Req1 = case Action of
         index ->
-            {ok, Ajax} = imboy_req:get_int(ajax, Req0, -2),
+            {ok, Ajax} = imboy_param:int(ajax, Req0, -2),
             % imboy_log:info(["AjaxAjaxAjaxAjaxAjax: ", Ajax, ";  Ajax"]),
             index(Method, Ajax, Req0, State);
         save ->
@@ -56,7 +56,7 @@ init(Req0, State0) ->
 %% @return cowboy_req:req() 更新后的请求对象
 -spec index(binary(), integer(), cowboy_req:req(), map()) -> cowboy_req:req().
 index(<<"GET">>, 1, Req0, _State) ->
-    {Page, Size} = imboy_req:page_size(Req0),
+    {Page, Size} = imboy_param:page(Req0),
     % Where = imboy_cnv:implode("", [<<"user_id=">>, CurrentUid]),
     % Where2 = <<"status > 0 AND ", Where/binary>>,
     Where = <<"1=1">>,